joystick touchevent not working

+++ src/com/mygame/WallCarver.java
@@ -3,14 +3,14 @@ package com.mygame;
 
 public class WallCarver{
 
-    protected byte _size;
+    protected int _size;
 	
-    protected byte _grid[][];
-    public WallCarver(byte size){
+    protected int _grid[][];
+    public WallCarver(int size){
     	
         _size = size;
         
-        _grid = new byte[_size][_size];
+        _grid = new int[_size][_size];
      
         for(int i=0; i<_size; i++){
         	for(int j=0; j<_size; j++)
